requirements.txt
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.23.5
omegaconf==2.3.0
pandas==1.4.*
pytorch_lifestream==0.5.2
pytorch_lightning==1.6.5
scikit_learn==1.2.1
torch==2.0.0
torchmetrics==0.11.2
wandb==0.15.10
